Allow retaining the original request Host header
Right now the `Host` header is always removed prior to making the target
request.  This commit allows you to alter the behavior using the
`config.headers.host` configuration.  The default value is `true`,
which means set (or reset) the `Host` header.  If set to `false`, it
means do not set (or reset) the `Host` header and the originally
requested `Host` header, if there is one, will be retained when making
the target request.