Add configurable keepAliveTimeout
This solves an error where a load balancer will try to use a connection
that node has closed. In GCP, this erorr shows up in the logs as:
backend_connection_closed_before_data_sent_to_client
Google's documentation states that their keepalive time is set to 10mins
and to try setting your server time to greater than that. This enables
microgateway to set that (beyond the 5sec default) and be configurable
based on a clients needs.

Details:
https://blog.percy.io/tuning-nginx-behind-google-cloud-platform-http-s-load-balancer-305982ddb340
https://cloud.google.com/load-balancing/docs/https/https-logging-monitoring